Add Spectrum as an Adobe Pass provider
Some services no longer allow for the older services (ex.
`Charter_Direct`) to be used and just return a generic 401 response.
Going through their login flow manually, the new provider name appears to
be `Spectrum`. This provider seems to work even place of services that
still allow for `Charter_Direct` to be used.

The Spectrum login form also needs special handling since it's
dynamically loaded via JS thus we cannot simply capture the POST url and
hidden form fields. Instead we need to hardcode the POST URL and
recreate the request manually. Thankfully the SAMLRequest and RelayState
are in the HTML response.